+++ tracer/src/emulator/device/virtio_block_disk.rs
@@ -90,7 +90,7 @@ impl VirtioBlockDisk {
     ///
     /// # Arguments
     /// * `memory`
-    pub fn tick<const XLEN: u8>(&mut self, memory: &mut MemoryWrapper<{XLEN}>) {
+    pub fn tick<const XLEN: usize>(&mut self, memory: &mut MemoryWrapper<{XLEN}>) {
         if !self.notify_clocks.is_empty()
             && (self.clock == self.notify_clocks[0] + DISK_ACCESS_DELAY)
         {
@@ -318,7 +318,7 @@ impl VirtioBlockDisk {
     /// * `mem_address` Physical address. Must be eight-byte aligned.
     /// * `disk_address` Must be eight-byte aligned.
     /// * `length` Must be eight-byte aligned.
-    fn transfer_from_disk<const XLEN: u8>(
+    fn transfer_from_disk<const XLEN: usize>(
         &mut self,
         memory: &mut MemoryWrapper<XLEN>,
         mem_address: u64,
@@ -353,7 +353,7 @@ impl VirtioBlockDisk {
     /// * `mem_address` Physical address. Must be eight-byte aligned.
     /// * `disk_address` Must be eight-byte aligned.
     /// * `length` Must be eight-byte aligned.
-    fn transfer_to_disk<const XLEN: u8>(
+    fn transfer_to_disk<const XLEN: usize>(
         &mut self,
         memory: &mut MemoryWrapper<{XLEN}>,
         mem_address: u64,
@@ -454,7 +454,7 @@ impl VirtioBlockDisk {
     }
 
     // @TODO: Follow the virtio block specification more properly.
-    fn handle_disk_access<const XLEN: u8>(&mut self, memory: &mut MemoryWrapper<XLEN>) {
+    fn handle_disk_access<const XLEN: usize>(&mut self, memory: &mut MemoryWrapper<XLEN>) {
         let base_desc_address = self.get_base_desc_address();
         let base_avail_address = self.get_base_avail_address();
         let base_used_address = self.get_base_used_address();
